refactor: refatorar validação de formulario da home
